Reset self.services when calling config()
I'm not sure what the semantics of collectd are -- whether a config load
always coincides with reinstantiating the plugin, or whether `config()` can
be called twice against the same object.

Nevertheless, I think it would be surprising to retain old services, if
loading a config file that contains a new set of services. So reset the dict
whenever calling `config()`.