Use requests HTTP client library
This is a lot simpler to use and will make the upcoming handling of cookies
much easier to implement. I think the reason I didn't use requests
originally is because we were installing the lone Python script onto a
server using Puppet and would have needed to describe the dependencies
separately, but now that we're using `pip` it will be handled for us.

NB: You'll need to delete your existing `.tox/` directory to pick up the new
dependencies due to this bug:

https://bitbucket.org/hpk42/tox/issue/93/environment-is-not-recreated-when