Cast INTERVAL to float during scale_and_type()
Casting INTERVAL to a float ensures that the divison returns a float
rather than an integer. In Python 3 division will operate like this
automatically, but for now we can force the desired behaviour.

This improves the accuracy of the data returned to collectd as noted in
cf50005f3bf67e1c9c60c8ad49e69dfdc198deb3.